Move IcebergRecordReader and IcebergSplit out of IcebergInputFormat
