+++ mr/src/test/java/org/apache/iceberg/mr/mapreduce/TestIcebergInputFormat.java
@@ -406,7 +406,7 @@ public class TestIcebergInputFormat {
     configBuilder.readFrom(location.toString());
 
     for (InputSplit split : splits(job.getConfiguration())) {
-      Assert.assertArrayEquals(IcebergInputFormat.IcebergSplit.ANYWHERE, split.getLocations());
+      Assert.assertArrayEquals(IcebergSplit.ANYWHERE, split.getLocations());
     }
 
     configBuilder.preferLocality();
